web/requirements.txt
alembic==1.0.8
aniso8601==6.0.0
bcrypt==3.1.6
cffi==1.12.2
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.7
Flask-JWT-Extended==3.18.0
Flask-Migrate==2.4.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SocketIO==3.3.2
Flask-SQLAlchemy==2.3.2
gevent==1.4.0
gunicorn==19.9.0
htmlmin==0.1.12
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.1
marshmallow==2.19.1
marshmallow-sqlalchemy==0.16.1
psutil==5.6.1
psycopg2==2.7.7
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
python-engineio==3.5.0
python-socketio==3.1.2
pytz==2018.9
six==1.12.0
SQLAlchemy==1.3.1
Werkzeug==0.14.1